+++ CHANGELOG.md
@@ -29,3 +29,10 @@ All notable changes to the "codeshare" extension will be documented in this file
 ### Fixed
 
 - Fixed spelling misstakes
+
+## [1.0.5] - 2020-11-29
+
+### Added
+
+- Added "Remove Share"-Function in Share-History
+- Added Icons in Context-Menu